backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.20

sqlalchemy==2.0.36
psycopg2-binary==2.9.10

python-dotenv==1.0.1
python-jose==3.3.0

passlib[argon2]==1.7.4
argon2-cffi==23.1.0

qrcode[pil]==7.4.2
pillow==10.4.0

openpyxl==3.1.5

tzdata==2025.2
